split into planer_common and rtc_interface
