+++ planning/behavior_velocity_planner/autoware_behavior_velocity_intersection_module/src/manager.hpp
@@ -21,6 +21,7 @@
 #include <autoware/behavior_velocity_planner_common/plugin_interface.hpp>
 #include <autoware/behavior_velocity_planner_common/plugin_wrapper.hpp>
 #include <autoware/behavior_velocity_planner_common/scene_module_interface.hpp>
+#include <autoware/behavior_velocity_rtc_interface/scene_module_interface_with_rtc.hpp>
 #include <rclcpp/rclcpp.hpp>
 
 #include <std_msgs/msg/string.hpp>
@@ -47,8 +48,8 @@ private:
 
   void launchNewModules(const tier4_planning_msgs::msg::PathWithLaneId & path) override;
 
-  std::function<bool(const std::shared_ptr<SceneModuleInterface> &)> getModuleExpiredFunction(
-    const tier4_planning_msgs::msg::PathWithLaneId & path) override;
+  std::function<bool(const std::shared_ptr<SceneModuleInterfaceWithRTC> &)>
+  getModuleExpiredFunction(const tier4_planning_msgs::msg::PathWithLaneId & path) override;
 
   bool hasSameParentLaneletAndTurnDirectionWithRegistered(const lanelet::ConstLanelet & lane) const;
 
@@ -63,7 +64,7 @@ private:
     tl_observation_pub_;
 };
 
-class MergeFromPrivateModuleManager : public SceneModuleManagerInterface
+class MergeFromPrivateModuleManager : public SceneModuleManagerInterface<>
 {
 public:
   explicit MergeFromPrivateModuleManager(rclcpp::Node & node);
