+++ planning/behavior_velocity_planner/autoware_behavior_velocity_no_stopping_area_module/src/manager.cpp
@@ -69,16 +69,17 @@ void NoStoppingAreaModuleManager::launchNewModules(
   }
 }
 
-std::function<bool(const std::shared_ptr<SceneModuleInterface> &)>
+std::function<bool(const std::shared_ptr<SceneModuleInterfaceWithRTC> &)>
 NoStoppingAreaModuleManager::getModuleExpiredFunction(
   const tier4_planning_msgs::msg::PathWithLaneId & path)
 {
   const auto no_stopping_area_id_set = planning_utils::getRegElemIdSetOnPath<NoStoppingArea>(
     path, planner_data_->route_handler_->getLaneletMapPtr(), planner_data_->current_odometry->pose);
 
-  return [no_stopping_area_id_set](const std::shared_ptr<SceneModuleInterface> & scene_module) {
-    return no_stopping_area_id_set.count(scene_module->getModuleId()) == 0;
-  };
+  return
+    [no_stopping_area_id_set](const std::shared_ptr<SceneModuleInterfaceWithRTC> & scene_module) {
+      return no_stopping_area_id_set.count(scene_module->getModuleId()) == 0;
+    };
 }
 
 }  // namespace autoware::behavior_velocity_planner
