+++ planning/behavior_velocity_planner/autoware_behavior_velocity_no_stopping_area_module/src/scene_no_stopping_area.hpp
@@ -17,9 +17,9 @@
 
 #include "utils.hpp"
 
-#include <autoware/behavior_velocity_planner_common/scene_module_interface.hpp>
 #include <autoware/behavior_velocity_planner_common/utilization/boost_geometry_helper.hpp>
 #include <autoware/behavior_velocity_planner_common/utilization/state_machine.hpp>
+#include <autoware/behavior_velocity_rtc_interface/scene_module_interface_with_rtc.hpp>
 #include <autoware_lanelet2_extension/regulatory_elements/no_stopping_area.hpp>
 #include <rclcpp/rclcpp.hpp>
 
@@ -31,7 +31,7 @@
 
 namespace autoware::behavior_velocity_planner
 {
-class NoStoppingAreaModule : public SceneModuleInterface
+class NoStoppingAreaModule : public SceneModuleInterfaceWithRTC
 {
 public:
   struct PlannerParam
