+++ planning/behavior_velocity_planner/autoware_behavior_velocity_occlusion_spot_module/src/manager.hpp
@@ -39,7 +39,7 @@
 
 namespace autoware::behavior_velocity_planner
 {
-class OcclusionSpotModuleManager : public SceneModuleManagerInterface
+class OcclusionSpotModuleManager : public SceneModuleManagerInterface<>
 {
 public:
   explicit OcclusionSpotModuleManager(rclcpp::Node & node);
