+++ planning/behavior_velocity_planner/autoware_behavior_velocity_planner_common/README.md
@@ -1,3 +1,3 @@
 # Behavior Velocity Planner Common
 
-This package provides common functions as a library, which are used in the `behavior_velocity_planner` node and modules.
+This package provides a behavior velocity interface without RTC, and common functions as a library, which are used in the `behavior_velocity_planner` node and modules.
