+++ planning/behavior_velocity_planner/autoware_behavior_velocity_planner_common/include/autoware/behavior_velocity_planner_common/scene_module_interface.hpp
@@ -16,12 +16,14 @@
 #define AUTOWARE__BEHAVIOR_VELOCITY_PLANNER_COMMON__SCENE_MODULE_INTERFACE_HPP_
 
 #include <autoware/behavior_velocity_planner_common/planner_data.hpp>
+#include <autoware/behavior_velocity_planner_common/utilization/util.hpp>
 #include <autoware/motion_utils/factor/velocity_factor_interface.hpp>
 #include <autoware/motion_utils/marker/virtual_wall_marker_creator.hpp>
+#include <autoware/motion_utils/trajectory/trajectory.hpp>
 #include <autoware/objects_of_interest_marker_interface/objects_of_interest_marker_interface.hpp>
-#include <autoware/rtc_interface/rtc_interface.hpp>
 #include <autoware/universe_utils/ros/debug_publisher.hpp>
 #include <autoware/universe_utils/ros/parameter.hpp>
+#include <autoware/universe_utils/system/stop_watch.hpp>
 #include <autoware/universe_utils/system/time_keeper.hpp>
 #include <builtin_interfaces/msg/time.hpp>
 
@@ -30,7 +32,6 @@
 #include <autoware_internal_debug_msgs/msg/float64_stamped.hpp>
 #include <autoware_planning_msgs/msg/path.hpp>
 #include <tier4_planning_msgs/msg/path_with_lane_id.hpp>
-#include <tier4_rtc_msgs/msg/state.hpp>
 #include <tier4_v2x_msgs/msg/infrastructure_command_array.hpp>
 #include <unique_identifier_msgs/msg/uuid.hpp>
 
@@ -54,14 +55,12 @@ using autoware::motion_utils::PlanningBehavior;
 using autoware::motion_utils::VelocityFactor;
 using autoware::objects_of_interest_marker_interface::ColorName;
 using autoware::objects_of_interest_marker_interface::ObjectsOfInterestMarkerInterface;
-using autoware::rtc_interface::RTCInterface;
 using autoware::universe_utils::DebugPublisher;
 using autoware::universe_utils::getOrDeclareParameter;
+using autoware::universe_utils::StopWatch;
 using autoware_internal_debug_msgs::msg::Float64Stamped;
 using builtin_interfaces::msg::Time;
 using tier4_planning_msgs::msg::PathWithLaneId;
-using tier4_rtc_msgs::msg::Module;
-using tier4_rtc_msgs::msg::State;
 using unique_identifier_msgs::msg::UUID;
 
 struct ObjectOfInterest
@@ -114,12 +113,6 @@ public:
 
   std::shared_ptr<universe_utils::TimeKeeper> getTimeKeeper() { return time_keeper_; }
 
-  void setActivation(const bool activated) { activated_ = activated; }
-  void setRTCEnabled(const bool enable_rtc) { rtc_enabled_ = enable_rtc; }
-  bool isActivated() const { return activated_; }
-  bool isSafe() const { return safe_; }
-  double getDistance() const { return distance_; }
-
   void resetVelocityFactor() { velocity_factor_.reset(); }
   VelocityFactor getVelocityFactor() const { return velocity_factor_.get(); }
   std::vector<ObjectOfInterest> getObjectsOfInterestData() const { return objects_of_interest_; }
@@ -127,10 +120,6 @@ public:
 
 protected:
   const int64_t module_id_;
-  bool activated_;
-  bool safe_;
-  bool rtc_enabled_;
-  double distance_;
   rclcpp::Logger logger_;
   rclcpp::Clock::SharedPtr clock_;
   std::shared_ptr<const PlannerData> planner_data_;
@@ -138,16 +127,6 @@ protected:
   autoware::motion_utils::VelocityFactorInterface velocity_factor_;
   std::vector<ObjectOfInterest> objects_of_interest_;
   mutable std::shared_ptr<universe_utils::TimeKeeper> time_keeper_;
-
-  void setSafe(const bool safe)
-  {
-    safe_ = safe;
-    if (!rtc_enabled_) {
-      syncActivation();
-    }
-  }
-  void setDistance(const double distance) { distance_ = distance; }
-  void syncActivation() { setActivation(isSafe()); }
 
   void setObjectsOfInterestData(
     const geometry_msgs::msg::Pose & pose, const autoware_perception_msgs::msg::Shape & shape,
@@ -160,10 +139,39 @@ protected:
     const std::vector<tier4_planning_msgs::msg::PathPointWithLaneId> & points) const;
 };
 
+template <class T = SceneModuleInterface>
 class SceneModuleManagerInterface
 {
 public:
-  SceneModuleManagerInterface(rclcpp::Node & node, [[maybe_unused]] const char * module_name);
+  SceneModuleManagerInterface(rclcpp::Node & node, [[maybe_unused]] const char * module_name)
+  : node_(node), clock_(node.get_clock()), logger_(node.get_logger())
+  {
+    const auto ns = std::string("~/debug/") + module_name;
+    pub_debug_ = node.create_publisher<visualization_msgs::msg::MarkerArray>(ns, 1);
+    if (!node.has_parameter("is_publish_debug_path")) {
+      is_publish_debug_path_ = node.declare_parameter<bool>("is_publish_debug_path");
+    } else {
+      is_publish_debug_path_ = node.get_parameter("is_publish_debug_path").as_bool();
+    }
+    if (is_publish_debug_path_) {
+      pub_debug_path_ = node.create_publisher<tier4_planning_msgs::msg::PathWithLaneId>(
+        std::string("~/debug/path_with_lane_id/") + module_name, 1);
+    }
+    pub_virtual_wall_ = node.create_publisher<visualization_msgs::msg::MarkerArray>(
+      std::string("~/virtual_wall/") + module_name, 5);
+    pub_velocity_factor_ = node.create_publisher<autoware_adapi_v1_msgs::msg::VelocityFactorArray>(
+      std::string("/planning/velocity_factors/") + module_name, 1);
+    pub_infrastructure_commands_ =
+      node.create_publisher<tier4_v2x_msgs::msg::InfrastructureCommandArray>(
+        "~/output/infrastructure_commands", 1);
+
+    processing_time_publisher_ = std::make_shared<DebugPublisher>(&node, "~/debug");
+
+    pub_processing_time_detail_ = node.create_publisher<universe_utils::ProcessingTimeDetail>(
+      "~/debug/processing_time_detail_ms/" + std::string(module_name), 1);
+
+    time_keeper_ = std::make_shared<universe_utils::TimeKeeper>(pub_processing_time_detail_);
+  }
 
   virtual ~SceneModuleManagerInterface() = default;
 
@@ -171,31 +179,111 @@ public:
 
   void updateSceneModuleInstances(
     const std::shared_ptr<const PlannerData> & planner_data,
-    const tier4_planning_msgs::msg::PathWithLaneId & path);
+    const tier4_planning_msgs::msg::PathWithLaneId & path)
+  {
+    planner_data_ = planner_data;
+
+    launchNewModules(path);
+    deleteExpiredModules(path);
+  }
 
   virtual void plan(tier4_planning_msgs::msg::PathWithLaneId * path) { modifyPathVelocity(path); }
 
 protected:
-  virtual void modifyPathVelocity(tier4_planning_msgs::msg::PathWithLaneId * path);
+  virtual void modifyPathVelocity(tier4_planning_msgs::msg::PathWithLaneId * path)
+  {
+    universe_utils::ScopedTimeTrack st(
+      "SceneModuleManagerInterface::modifyPathVelocity", *time_keeper_);
+    StopWatch<std::chrono::milliseconds> stop_watch;
+    stop_watch.tic("Total");
+    visualization_msgs::msg::MarkerArray debug_marker_array;
+    autoware_adapi_v1_msgs::msg::VelocityFactorArray velocity_factor_array;
+    velocity_factor_array.header.frame_id = "map";
+    velocity_factor_array.header.stamp = clock_->now();
+
+    tier4_v2x_msgs::msg::InfrastructureCommandArray infrastructure_command_array;
+    infrastructure_command_array.stamp = clock_->now();
+
+    for (const auto & scene_module : scene_modules_) {
+      scene_module->resetVelocityFactor();
+      scene_module->setPlannerData(planner_data_);
+      scene_module->modifyPathVelocity(path);
+
+      // The velocity factor must be called after modifyPathVelocity.
+      const auto velocity_factor = scene_module->getVelocityFactor();
+      if (velocity_factor.behavior != PlanningBehavior::UNKNOWN) {
+        velocity_factor_array.factors.emplace_back(velocity_factor);
+      }
+
+      if (const auto command = scene_module->getInfrastructureCommand()) {
+        infrastructure_command_array.commands.push_back(*command);
+      }
+
+      for (const auto & marker : scene_module->createDebugMarkerArray().markers) {
+        debug_marker_array.markers.push_back(marker);
+      }
+
+      virtual_wall_marker_creator_.add_virtual_walls(scene_module->createVirtualWalls());
+    }
+
+    pub_velocity_factor_->publish(velocity_factor_array);
+    pub_infrastructure_commands_->publish(infrastructure_command_array);
+    pub_debug_->publish(debug_marker_array);
+    if (is_publish_debug_path_) {
+      tier4_planning_msgs::msg::PathWithLaneId debug_path;
+      debug_path.header = path->header;
+      debug_path.points = path->points;
+      pub_debug_path_->publish(debug_path);
+    }
+    pub_virtual_wall_->publish(virtual_wall_marker_creator_.create_markers(clock_->now()));
+    processing_time_publisher_->publish<Float64Stamped>(
+      std::string(getModuleName()) + "/processing_time_ms", stop_watch.toc("Total"));
+  }
 
   virtual void launchNewModules(const tier4_planning_msgs::msg::PathWithLaneId & path) = 0;
 
-  virtual std::function<bool(const std::shared_ptr<SceneModuleInterface> &)>
-  getModuleExpiredFunction(const tier4_planning_msgs::msg::PathWithLaneId & path) = 0;
+  virtual std::function<bool(const std::shared_ptr<T> &)> getModuleExpiredFunction(
+    const tier4_planning_msgs::msg::PathWithLaneId & path) = 0;
 
-  virtual void deleteExpiredModules(const tier4_planning_msgs::msg::PathWithLaneId & path);
+  virtual void deleteExpiredModules(const tier4_planning_msgs::msg::PathWithLaneId & path)
+  {
+    const auto isModuleExpired = getModuleExpiredFunction(path);
+
+    auto itr = scene_modules_.begin();
+    while (itr != scene_modules_.end()) {
+      if (isModuleExpired(*itr)) {
+        itr = scene_modules_.erase(itr);
+      } else {
+        itr++;
+      }
+    }
+  }
 
   bool isModuleRegistered(const int64_t module_id)
   {
     return registered_module_id_set_.count(module_id) != 0;
   }
 
-  void registerModule(const std::shared_ptr<SceneModuleInterface> & scene_module);
+  void registerModule(const std::shared_ptr<T> & scene_module)
+  {
+    RCLCPP_DEBUG(
+      logger_, "register task: module = %s, id = %lu", getModuleName(),
+      scene_module->getModuleId());
+    registered_module_id_set_.emplace(scene_module->getModuleId());
+    scene_module->setTimeKeeper(time_keeper_);
+    scene_modules_.insert(scene_module);
+  }
 
   size_t findEgoSegmentIndex(
-    const std::vector<tier4_planning_msgs::msg::PathPointWithLaneId> & points) const;
+    const std::vector<tier4_planning_msgs::msg::PathPointWithLaneId> & points) const
+  {
+    const auto & p = planner_data_;
+    return autoware::motion_utils::findFirstNearestSegmentIndexWithSoftConstraints(
+      points, p->current_odometry->pose, p->ego_nearest_dist_threshold,
+      p->ego_nearest_yaw_threshold);
+  }
 
-  std::set<std::shared_ptr<SceneModuleInterface>> scene_modules_;
+  std::set<std::shared_ptr<T>> scene_modules_;
   std::set<int64_t> registered_module_id_set_;
 
   std::shared_ptr<const PlannerData> planner_data_;
@@ -220,66 +308,6 @@ protected:
 
   std::shared_ptr<universe_utils::TimeKeeper> time_keeper_;
 };
-
-class SceneModuleManagerInterfaceWithRTC : public SceneModuleManagerInterface
-{
-public:
-  SceneModuleManagerInterfaceWithRTC(
-    rclcpp::Node & node, const char * module_name, const bool enable_rtc = true);
-
-  void plan(tier4_planning_msgs::msg::PathWithLaneId * path) override;
-
-protected:
-  RTCInterface rtc_interface_;
-  std::unordered_map<int64_t, UUID> map_uuid_;
-
-  ObjectsOfInterestMarkerInterface objects_of_interest_marker_interface_;
-
-  virtual void sendRTC(const Time & stamp);
-
-  virtual void setActivation();
-
-  void updateRTCStatus(
-    const UUID & uuid, const bool safe, const uint8_t state, const double distance,
-    const Time & stamp)
-  {
-    rtc_interface_.updateCooperateStatus(uuid, safe, state, distance, distance, stamp);
-  }
-
-  void removeRTCStatus(const UUID & uuid) { rtc_interface_.removeCooperateStatus(uuid); }
-
-  void publishRTCStatus(const Time & stamp)
-  {
-    rtc_interface_.removeExpiredCooperateStatus();
-    rtc_interface_.publishCooperateStatus(stamp);
-  }
-
-  UUID getUUID(const int64_t & module_id) const;
-
-  void generateUUID(const int64_t & module_id);
-
-  void removeUUID(const int64_t & module_id);
-
-  void publishObjectsOfInterestMarker();
-
-  void deleteExpiredModules(const tier4_planning_msgs::msg::PathWithLaneId & path) override;
-
-  static bool getEnableRTC(rclcpp::Node & node, const std::string & param_name)
-  {
-    bool enable_rtc = true;
-
-    try {
-      enable_rtc = getOrDeclareParameter<bool>(node, "enable_all_modules_auto_mode")
-                     ? false
-                     : getOrDeclareParameter<bool>(node, param_name);
-    } catch (const std::exception & e) {
-      enable_rtc = getOrDeclareParameter<bool>(node, param_name);
-    }
-
-    return enable_rtc;
-  }
-};
-
 }  // namespace autoware::behavior_velocity_planner
 
 #endif  // AUTOWARE__BEHAVIOR_VELOCITY_PLANNER_COMMON__SCENE_MODULE_INTERFACE_HPP_
