+++ planning/behavior_velocity_planner/autoware_behavior_velocity_rtc_interface/README.md
@@ -1,0 +1,3 @@
+# Behavior Velocity RTC Interface
+
+This package provides a behavior velocity interface with RTC, which are used in the `behavior_velocity_planner` node and modules.
