+++ planning/behavior_velocity_planner/autoware_behavior_velocity_rtc_interface/package.xml
@@ -1,47 +1,32 @@
 <?xml version="1.0"?>
 <?xml-model href="http://download.ros.org/schema/package_format3.xsd" schematypens="http://www.w3.org/2001/XMLSchema"?>
 <package format="3">
-  <name>autoware_behavior_velocity_intersection_module</name>
+  <name>autoware_behavior_velocity_rtc_interface</name>
   <version>0.40.0</version>
-  <description>The autoware_behavior_velocity_intersection_module package</description>
+  <description>The autoware_behavior_velocity_rtc_interface package</description>
 
   <maintainer email="mamoru.sobue@tier4.jp">Mamoru Sobue</maintainer>
   <maintainer email="takayuki.murooka@tier4.jp">Takayuki Murooka</maintainer>
   <maintainer email="tomoya.kimura@tier4.jp">Tomoya Kimura</maintainer>
   <maintainer email="shumpei.wakabayashi@tier4.jp">Shumpei Wakabayashi</maintainer>
-  <maintainer email="kyoichi.sugahara@tier4.jp">Kyoichi Sugahara</maintainer>
-  <maintainer email="yukinari.hisaki.2@tier4.jp">Yukinari Hisaki</maintainer>
+  <maintainer email="isamu.takagi@tier4.jp">Takagi, Isamu</maintainer>
+  <maintainer email="fumiya.watanabe@tier4.jp">Fumiya Watanabe</maintainer>
 
   <license>Apache License 2.0</license>
 
-  <author email="mamoru.sobue@tier4.jp">Mamoru Sobue</author>
+  <author email="takayuki.murooka@tier4.jp">Takayuki Murooka</author>
 
   <buildtool_depend>ament_cmake_auto</buildtool_depend>
   <buildtool_depend>autoware_cmake</buildtool_depend>
 
   <depend>autoware_behavior_velocity_planner_common</depend>
-  <depend>autoware_behavior_velocity_rtc_interface</depend>
-  <depend>autoware_internal_debug_msgs</depend>
-  <depend>autoware_interpolation</depend>
-  <depend>autoware_lanelet2_extension</depend>
   <depend>autoware_motion_utils</depend>
-  <depend>autoware_perception_msgs</depend>
   <depend>autoware_planning_msgs</depend>
-  <depend>autoware_route_handler</depend>
   <depend>autoware_rtc_interface</depend>
-  <depend>autoware_test_utils</depend>
   <depend>autoware_universe_utils</depend>
-  <depend>autoware_vehicle_info_utils</depend>
-  <depend>fmt</depend>
-  <depend>geometry_msgs</depend>
-  <depend>libopencv-dev</depend>
-  <depend>magic_enum</depend>
-  <depend>nav_msgs</depend>
-  <depend>pluginlib</depend>
   <depend>rclcpp</depend>
-  <depend>tf2_geometry_msgs</depend>
+  <depend>rclcpp_components</depend>
   <depend>tier4_planning_msgs</depend>
-  <depend>visualization_msgs</depend>
 
   <test_depend>ament_cmake_ros</test_depend>
   <test_depend>ament_lint_auto</test_depend>
