+++ planning/behavior_velocity_planner/autoware_behavior_velocity_speed_bump_module/src/manager.hpp
@@ -29,7 +29,7 @@
 
 namespace autoware::behavior_velocity_planner
 {
-class SpeedBumpModuleManager : public SceneModuleManagerInterface
+class SpeedBumpModuleManager : public SceneModuleManagerInterface<>
 {
 public:
   explicit SpeedBumpModuleManager(rclcpp::Node & node);
