+++ planning/behavior_velocity_planner/autoware_behavior_velocity_traffic_light_module/package.xml
@@ -20,6 +20,7 @@
 
   <depend>autoware_adapi_v1_msgs</depend>
   <depend>autoware_behavior_velocity_planner_common</depend>
+  <depend>autoware_behavior_velocity_rtc_interface</depend>
   <depend>autoware_lanelet2_extension</depend>
   <depend>autoware_motion_utils</depend>
   <depend>autoware_perception_msgs</depend>
