+++ planning/behavior_velocity_planner/autoware_behavior_velocity_traffic_light_module/src/scene.hpp
@@ -23,8 +23,8 @@
 #define EIGEN_MPL2_ONLY
 #include <Eigen/Core>
 #include <Eigen/Geometry>
-#include <autoware/behavior_velocity_planner_common/scene_module_interface.hpp>
 #include <autoware/behavior_velocity_planner_common/utilization/boost_geometry_helper.hpp>
+#include <autoware/behavior_velocity_rtc_interface/scene_module_interface_with_rtc.hpp>
 #include <autoware_lanelet2_extension/utility/query.hpp>
 #include <rclcpp/rclcpp.hpp>
 
@@ -33,7 +33,7 @@
 
 namespace autoware::behavior_velocity_planner
 {
-class TrafficLightModule : public SceneModuleInterface
+class TrafficLightModule : public SceneModuleInterfaceWithRTC
 {
 public:
   using TrafficSignal = autoware_perception_msgs::msg::TrafficLightGroup;
